+++ src/com/normanwhittlecliff/cmi/states/GameStateAbout.java
@@ -1,0 +1,5 @@
+package com.normanwhittlecliff.cmi.states;
+
+public class GameStateAbout {
+
+}