+++ src/com/normanwhittlecliff/coffeeLib/core/GameWindow.java
@@ -8,13 +8,13 @@ public class GameWindow extends JFrame {
 	private static final long serialVersionUID = 1L;
 	//private Game game;
 	// =================================================================================================================
-	public GameWindow (Game game) {
+	public GameWindow (GameCore gameCore) {
 		//this.game = game;
 		
-		this.setTitle(game.getGameConfig().getTitle() + " (v" + game.getGameConfig().getVersion() + ")") ;
-		this.setResizable(game.getGameConfig().isWindowRezisable());
+		this.setTitle(gameCore.getGameConfig().getWindowTitle()) ;
+		this.setResizable(gameCore.getGameConfig().isWindowRezisable());
 		
-		this.add(game.getGamePanel());
+		this.add(gameCore.getGamePanel());
 		
 		this.setDefaultCloseOperation(JFrame.EXIT_ON_CLOSE);
 		this.pack();