+++ src/com/normanwhittlecliff/coffeeLib/state/GameStateMethods.java
@@ -1,0 +1,15 @@
+package com.normanwhittlecliff.coffeeLib.state;
+
+import java.awt.Graphics;
+
+public interface GameStateMethods {
+	
+	public void init();
+	
+	public void update();
+	
+	public void render(Graphics g);
+	
+	public void inputs();
+	
+}