+++ CHANGELOG.md
@@ -6,7 +6,15 @@ All notable changes to this project will be documented in this file.
 
 ---
 
-## [0.0.0 alpha] - 2026-03-25
+## [0.0.0 pre-alpha] - 2026-03-24
+### Added
+- Project idea definition
+- Core concept: coffee-making story game
+
+---
+
+## [0.0.0 pre-alpha] - 2026-03-25
+
 ### Added
 - Initial project structure
 - Basic folder organization
@@ -18,7 +26,24 @@ All notable changes to this project will be documented in this file.
 
 ---
 
-## [0.0.1] - 2026-03-24
-### Added
-- Project idea definition
-- Core concept: coffee-making story game
+## [0.0.0 pre-alpha] - 2026-03-29
+
+### Added (Engine)
+- Input dedicated class, but it's not working with a functional gameloop and a dedicated panel
+- Window dedicated class
+- Panel dedicated class
+- GameLoop dedicated class
+
+---
+
+## [0.0.0 pre-alpha] - 2026-03-30
+
+---
+
+## [0.0.0 pre-alpha] - 2026-04-02
+
+### Created
+- Timer class
+
+### Finished (Engine)
+- Input class. Easy to add to any class in the Game now.